handle messages immutably in reducers
